Removed unnecessary documents
Previously, this repository was used as a dumping place for all
sorts of literature that I found interesting and plan to read. I
decided that this has wandered from the original purpose of this
repo, which is to contain documents that describe the code contained
in my other repositories. Thus, I have removed all unnecessary
documents.