Add what-makes-a-kernel.txt, add to features.txt
Neither of these is finished yet. `features.txt' is the file that
I'm using to store ideas for my microcontroller/kernel/filesystem
project. `what-makes-a-kernel.txt' is a list of all the kinds of
interfaces that a kernel is expected to export. Be warned, some of
it is Linux specific, and some may be biased to monolithic kernels.